% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_functions.R
\name{AREMP_MMI_model}
\alias{AREMP_MMI_model}
\title{AREMP MMI}
\usage{
AREMP_MMI_model(
  bugnew,
  prednew,
  CLING_rich.rf,
  DIPT_rich.rf,
  LLT_rich.rf,
  NON_INSECT_rich.rf,
  PER_EPT.rf,
  PER_INTOL.rf,
  rf_models,
  mdeg_metrics_adj_cal,
  ref_metrics_adj
)
}
\arguments{
\item{bugnew}{}

\item{prednew}{}

\item{CLING_rich.rf}{}

\item{DIPT_rich.rf}{}

\item{LLT_rich.rf}{}

\item{NON_INSECT_rich.rf}{}

\item{PER_EPT.rf}{}

\item{PER_INTOL.rf}{}

\item{rf_models}{}

\item{mdeg_metrics_adj_cal}{}

\item{ref_metrics_adj}{}
}
\value{
MMI value
}
\description{
AREMP MMI
}
