% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bug_functions.R
\name{CO_bug_export}
\alias{CO_bug_export}
\title{CO EDAS 2017 bug data export}
\usage{
CO_bug_export(boxId, translationId)
}
\arguments{
\item{boxId}{}

\item{translationId}{}
}
\value{
raw bug data translated by CO EDAS translation and with column names formatted for CO EDAS access database
}
\description{
CO EDAS 2017 bug data export
}
