% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bug_functions.R
\name{CSCI_bug}
\alias{CSCI_bug}
\title{CA CSCI bug data export}
\usage{
CSCI_bug(sampleId, translationId)
}
\arguments{
\item{sampleId}{}

\item{translationId}{}
}
\value{
raw bug data translated by CSCI translation and with column names formatted for CSCI model function
}
\description{
CA CSCI bug data export
}
