% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_functions.R
\name{metricMatrixRescale}
\alias{metricMatrixRescale}
\title{MMI metric rescaling}
\usage{
metricMatrixRescale(metrics, ref_metrics, mostdeg_metrics)
}
\arguments{
\item{metrics}{}

\item{ref_metrics}{}

\item{mostdeg_metrics}{}
}
\value{
rescaled standardized metrics to min and max of reference and most degraded data
}
\description{
MMI metric rescaling
}
