% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bug_functions.R
\name{MMI_metrics}
\alias{MMI_metrics}
\title{MMI metrics}
\usage{
MMI_metrics(sampleId, fixedCount, modelId)
}
\arguments{
\item{sampleId}{}

\item{fixedCount}{}
}
\value{
sample by metrics dataframe, only metrics needed for a given model
}
\description{
MMI metrics
}
