% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.predict.RanFor.4.2.r
\name{model.predict.RanFor.4.2}
\alias{model.predict.RanFor.4.2}
\title{OE predict 4.2}
\usage{
model.predict.RanFor.4.2(
  bugcal.pa,
  grps.final,
  preds.final,
  ranfor.mod,
  prednew,
  bugnew,
  Pc,
  Cal.OOB = FALSE
)
}
\arguments{
\item{bugcal.pa}{}

\item{grps.final}{}

\item{preds.final}{}

\item{ranfor.mod}{}

\item{prednew}{}

\item{bugnew}{}

\item{Pc}{}

\item{Cal.OOB}{}
}
\value{
O, E, OoverE, BC, null
}
\description{
OE predict 4.2
}
