% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.predict.RanFor.r
\name{model.predict.RanFor}
\alias{model.predict.RanFor}
\title{OE predict version 0}
\usage{
model.predict.RanFor(
  bugcal.pa,
  grps.final,
  preds.final,
  ranfor.mod,
  prednew,
  bugnew.pa,
  Pc = 0.5
)
}
\arguments{
\item{bugcal.pa}{}

\item{grps.final}{}

\item{preds.final}{}

\item{ranfor.mod}{}

\item{prednew}{}

\item{bugnew.pa}{}

\item{Pc}{}
}
\value{
O, E, OoverE, BC, null
}
\description{
OE predict version 0
}
