% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.predict.v4.1.r
\name{model.predict.v4.1}
\alias{model.predict.v4.1}
\title{OE predict version 4.1}
\usage{
model.predict.v4.1(
  bugcal.pa,
  grps.final,
  preds.final,
  grpmns,
  covpinv,
  prednew,
  bugnew,
  Pc
)
}
\arguments{
\item{bugcal.pa}{}

\item{grps.final}{}

\item{preds.final}{}

\item{grpmns}{}

\item{covpinv}{}

\item{prednew}{}

\item{bugnew}{}

\item{Pc}{}
}
\value{
O, E, OoverE, BC, null
}
\description{
OE predict version 4.1
}
