% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_functions.R
\name{NV_MMI_model}
\alias{NV_MMI_model}
\title{NV MMI}
\usage{
NV_MMI_model(
  bugnew,
  prednew,
  CLINGER.rf,
  INSET.rf,
  NONSET.rf,
  PER_CFA.rf,
  PER_EPHEA.rf,
  PER_PLECA.rf
)
}
\arguments{
\item{bugnew}{}

\item{prednew}{}

\item{CLINGER.rf}{}

\item{INSET.rf}{}

\item{NONSET.rf}{}

\item{PER_CFA.rf}{}

\item{PER_EPHEA.rf}{}

\item{PER_PLECA.rf}{}
}
\value{
MMI score
}
\description{
NV MMI
}
