% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bug_functions.R
\name{OE_bug_matrix}
\alias{OE_bug_matrix}
\title{OE bug data matrix export}
\usage{
OE_bug_matrix(sampleId, translationId, fixedCount)
}
\arguments{
\item{sampleId}{}

\item{translationId}{}

\item{fixedCount}{}
}
\value{
bug data sample by taxa presence/absence matrix
}
\description{
OE bug data matrix export
}
