% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bug_functions.R
\name{OR_NBR_bug}
\alias{OR_NBR_bug}
\title{OR NBR eastern}
\usage{
OR_NBR_bug(sampleId, translationId, fixedCount)
}
\arguments{
\item{sampleId}{}

\item{translationId}{}

\item{fixedCount}{}
}
\value{
list of rarefied and translated taxa per sample
}
\description{
OR NBR eastern
}
