% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_functions.R
\name{PIBO_model}
\alias{PIBO_model}
\title{PIBO model using version 0 of John Vansickles code}
\usage{
PIBO_model(
  bug.otu,
  bugall,
  bugnew,
  grps.final,
  preds.final,
  ranfor.mod,
  prednew
)
}
\arguments{
\item{bug.otu}{}

\item{bugall}{}

\item{bugnew}{}

\item{grps.final}{}

\item{preds.final}{}

\item{ranfor.mod}{}

\item{prednew}{}
}
\value{
O, E, OoverE,BC, null model info
}
\description{
PIBO model using version 0 of John Vansickles code
}
