% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_models.R
\name{process_sample_models}
\alias{process_sample_models}
\title{Process sample models}
\usage{
process_sample_models(sampleId, modelId, config = config)
}
\arguments{
\item{config}{}
}
\value{
none
}
\description{
Process sample models
}
\details{
saving each predictor for each sample one at a time in the database
}
